# docs/swagger.yaml
basePath: /
definitions:
  httpserver.CookieData:
    properties:
      mojavi:
        default: UrdBjDansj/s95fYW58TfQ
        type: string
        x-order: "02"
      siakng_cc:
        default: 2jNeTbVCFfkPIcnUkzwrVw
        type: string
        x-order: "01"
    type: object
  httpserver.LoginRequest:
    properties:
      password:
        default: password
        type: string
        x-order: "02"
      username:
        default: username
        type: string
        x-order: "01"
    type: object
  httpserver.Response-any:
    properties:
      data:
        x-order: "03"
      message:
        default: Success
        type: string
        x-order: "02"
      status:
        default: 200
        type: integer
        x-order: "01"
    type: object
...
  httpserver.Response-httpserver_CookieData:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/httpserver.CookieData'
        x-order: "03"
      message:
        default: Success
        type: string
        x-order: "02"
      status:
        default: 200
        type: integer
        x-order: "01"
    type: object
...
  httpserver.Response-siaklib_UserInfo:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/siaklib.UserInfo'
        x-order: "03"
      message:
        default: Success
        type: string
        x-order: "02"
      status:
        default: 200
        type: integer
        x-order: "01"
    type: object
...
  siaklib.UserInfo:
    properties:
      group:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
...
info:
  contact: {}
  description: This an REST API for siak
  title: Siak REST API
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: get login cookie for other requests
      parameters:
      - description: login request
        in: body
        name: loginRequest
        required: true
        schema:
          $ref: '#/definitions/httpserver.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpserver.Response-httpserver_CookieData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpserver.Response-any'
      summary: login account
  /me:
    get:
      description: get user account info
      parameters:
      - description: siakng cookie
        in: header
        name: X-Siakng-Cc
        required: true
        type: string
      - description: mojavi cookie
        in: header
        name: X-Mojavi
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpserver.Response-siaklib_UserInfo'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpserver.Response-any'
      summary: user account
swagger: "2.0"